requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
seqeval>=1.2.2
mlflow>=2.3.0
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
spacy>=3.5.0
tqdm>=4.65.0
pyyaml>=6.0

# PDF Processing
pdfplumber>=0.9.0
PyMuPDF>=1.22.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
python-multipart>=0.0.6
pytorch-crf>=0.7.2 # Provides 'torchcrf'
reportlab>=4.0.0
altair>=4.0.0

